# docs/swagger.yaml
definitions:
  dto.Article:
    properties:
      access:
        type: string
      alphabet_count:
        type: integer
      approved_at:
        type: string
      approved_by:
        type: string
      article_groups:
        items:
          $ref: '#/definitions/dto.ArticleGroup'
        type: array
      article_id_old:
        type: integer
      article_user:
        items:
          $ref: '#/definitions/dto.ArticleUser'
        type: array
      article_uuid:
        type: string
      attachments:
        items:
          $ref: '#/definitions/dto.Attachment'
        type: array
      breaking_news_at:
        type: string
      canonical_url:
        type: string
      canonical_url_old:
        type: string
      content:
        type: string
      content_category:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      domain_id:
        type: integer
      domain_id_old:
        type: integer
      embed_code:
        type: string
      fact_check_claim:
        type: string
      fact_check_result:
        type: string
      fact_check_result_image:
        type: string
      feature_image:
        type: string
      feature_image_caption:
        type: string
      foot_note:
        type: string
      format_article_id:
        type: integer
      headline_at:
        type: string
      id:
        type: integer
      infografis_file:
        type: string
      is_active:
        type: boolean
      is_anonymous_reporter_editor:
        type: boolean
      is_article_jembatan:
        type: string
      is_audio_available:
        type: boolean
      is_byline:
        type: boolean
      is_label_tag:
        description: will be deleted soon
        type: string
      kicker:
        type: string
      language:
        type: string
      paragraphs:
        items:
          type: string
        type: array
      published_at:
        type: string
      published_by:
        type: string
      snackbar:
        type: string
      status:
        type: string
      sub_rubric:
        $ref: '#/definitions/dto.SubRubric'
      sub_rubric_id:
        type: integer
      summary:
        items:
          type: string
        type: array
      tag_article:
        items:
          $ref: '#/definitions/dto.ArticleTagsOld'
        type: array
      tag_article_new:
        items:
          type: string
        type: array
      tag_title:
        items:
          type: string
        type: array
      taicing:
        type: string
      title_digital:
        type: string
      title_print:
        type: string
      title_seo:
        type: string
      tts_audio_url:
        type: string
      unpublished_at:
        type: string
      unpublished_by:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      upperdeck:
        type: string
      word_count:
        type: integer
      writing_style:
        type: string
    type: object
  dto.ArticleCreateRequestDto:
    properties:
      body:
        minLength: 5
        type: string
      title:
        minLength: 5
        type: string
    required:
    - body
    - title
    type: object
  dto.ArticleDetailResponse:
    properties:
      access:
        type: string
      alphabet_count:
        type: integer
      approved_at:
        type: string
      approved_by:
        type: string
      article_groups:
        items:
          $ref: '#/definitions/dto.ArticleGroup'
        type: array
      article_id_old:
        type: integer
      article_user:
        items:
          $ref: '#/definitions/dto.ArticleUser'
        type: array
      article_uuid:
        type: string
      attachments:
        items:
          $ref: '#/definitions/dto.Attachment'
        type: array
      breaking_news_at:
        type: string
      canonical_url:
        type: string
      canonical_url_old:
        type: string
      content:
        type: string
      content_category:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      domain_id:
        type: integer
      domain_id_old:
        type: integer
      embed_code:
        type: string
      fact_check_claim:
        type: string
      fact_check_result:
        type: string
      fact_check_result_image:
        type: string
      feature_image:
        type: string
      feature_image_caption:
        type: string
      foot_note:
        type: string
      format_article_id:
        type: integer
      headline_at:
        type: string
      id:
        type: integer
      infografis_file:
        type: string
      is_active:
        type: boolean
      is_anonymous_reporter_editor:
        type: boolean
      is_article_jembatan:
        type: string
      is_audio_available:
        type: boolean
      is_byline:
        type: boolean
      is_label_tag:
        description: will be deleted soon
        type: string
      kicker:
        type: string
      language:
        type: string
      paragraphs:
        items:
          type: string
        type: array
      published_at:
        type: string
      published_by:
        type: string
      snackbar:
        type: string
      status:
        type: string
      sub_rubric:
        $ref: '#/definitions/dto.SubRubric'
      sub_rubric_id:
        type: integer
      summary:
        items:
          type: string
        type: array
      tag_article:
        items:
          $ref: '#/definitions/dto.ArticleTagsOld'
        type: array
      tag_article_new:
        items:
          type: string
        type: array
      tag_title:
        items:
          type: string
        type: array
      taicing:
        type: string
      title_digital:
        type: string
      title_print:
        type: string
      title_seo:
        type: string
      tts_audio_url:
        type: string
      unpublished_at:
        type: string
      unpublished_by:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      upperdeck:
        type: string
      word_count:
        type: integer
      writing_style:
        type: string
    type: object
  dto.ArticleGroup:
    properties:
      article:
        $ref: '#/definitions/dto.Article'
      article_id:
        type: integer
      breaking_news_at:
        type: string
      breaking_news_by:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      group:
        $ref: '#/definitions/dto.Group'
      group_id:
        type: integer
      headline_at:
        type: string
      headline_by:
        type: string
      id:
        type: integer
      sequence_article_group:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  dto.ArticleTagsOld:
    properties:
      alias:
        type: string
      caption:
        type: string
    type: object
  dto.ArticleUpdateRequestDto:
    properties:
      body:
        minLength: 5
        type: string
      title:
        minLength: 5
        type: string
    required:
    - body
    - title
    type: object
  dto.ArticleUser:
    properties:
      article:
        $ref: '#/definitions/dto.Article'
      article_id:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      deleted_by:
        type: string
      id:
        type: integer
      type:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      user:
        $ref: '#/definitions/dto.User'
      user_id:
        type: integer
    type: object
  dto.Attachment:
    properties:
      articles:
        items:
          $ref: '#/definitions/dto.Article'
        type: array
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      deleted_by:
        type: string
      description:
        type: string
      editor:
        type: string
      id:
        type: integer
      is_video:
        type: boolean
      link:
        type: string
      name:
        type: string
      published_at:
        type: string
      published_by:
        type: string
      title:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      uuid:
        type: string
      video_source:
        type: string
      video_target:
        type: string
    type: object
  dto.AuthRequestDto:
    properties:
      email:
        minLength: 1
        type: string
      password:
        minLength: 5
        type: string
    required:
    - email
    - password
    type: object
  dto.Group:
    properties:
      alias:
        type: string
      article_groups:
        items:
          $ref: '#/definitions/dto.ArticleGroup'
        type: array
      articles:
        items:
          $ref: '#/definitions/dto.Article'
        type: array
      caption:
        type: string
      category:
        type: string
      cover:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      date:
        type: string
      deleted_at:
        type: string
      deleted_by:
        type: string
      description:
        type: string
      groups_child:
        items:
          $ref: '#/definitions/dto.Group'
        type: array
      id:
        type: integer
      is_active:
        type: boolean
      level:
        type: integer
      meta_keyword:
        type: string
      meta_title:
        type: string
      parent_id:
        type: integer
      published_at:
        type: string
      published_by:
        type: string
      sequence:
        type: integer
      status:
        type: string
      unpublished_at:
        type: string
      unpublished_by:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      uuid:
        type: string
    type: object
  dto.PaginationResponseDtoExample:
    properties:
      is_has_next_page:
        example: true
        type: boolean
      limit:
        type: integer
      list:
        items: {}
        type: array
      total_entry:
        example: 100
        type: integer
      total_page:
        type: integer
    type: object
  dto.Rubric:
    properties:
      alias:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      deleted_by:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      is_main:
        type: boolean
      name:
        type: string
      sub_rubric:
        items:
          $ref: '#/definitions/dto.SubRubric'
        type: array
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  dto.SubRubric:
    properties:
      alias:
        type: string
      articles:
        items:
          $ref: '#/definitions/dto.Article'
        type: array
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      deleted_by:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      rubric:
        $ref: '#/definitions/dto.Rubric'
      rubric_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  dto.User:
    properties:
      actived_at:
        type: string
      actived_by:
        type: string
      alias:
        type: string
      biodata:
        type: string
      biodata_en:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      deleted_by:
        type: string
      email:
        type: string
      facebook:
        type: string
      force_reset_password:
        type: boolean
      foto:
        type: string
      id:
        type: integer
      is_admin:
        type: boolean
      is_contributor:
        type: boolean
      is_editor:
        type: boolean
      linked:
        type: string
      main_email:
        type: string
      name:
        type: string
      nik:
        type: string
      password:
        type: string
      secondary_email:
        type: string
      twitter:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      user_remp_id:
        type: integer
      username:
        type: string
      uuid:
        type: string
    type: object
  dto.UserInfoResponse:
    properties:
      email:
        type: string
      foto:
        type: string
      id:
        type: integer
      name:
        type: string
      permits:
        items:
          type: string
        type: array
      role:
        type: string
      username:
        type: string
    type: object
  helper.Response:
    properties:
      data: {}
      message:
        example: ok
        type: string
      status:
        example: 200
        type: integer
    type: object
info:
  contact: {}
  title: Go Pustaka API Documentation
  version: "3.0"
paths:
  /client/articles:
    get:
      parameters:
      - example: 10
        in: query
        name: limit
        required: true
        type: integer
      - example: 1
        in: query
        name: page
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PaginationResponseDtoExample'
              type: object
      security:
      - BearerAuth: []
      summary: Get All Article
      tags:
      - Article
  /client/articles/{id}:
    get:
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ArticleDetailResponse'
              type: object
      security:
      - XApiKey: []
      - XTempoToken: []
      - XRequestTime: []
      summary: Get Article Details
      tags:
      - Article
  /newsroom/articles:
    get:
      parameters:
      - example: 10
        in: query
        name: limit
        required: true
        type: integer
      - example: 1
        in: query
        name: page
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PaginationResponseDtoExample'
              type: object
      security:
      - BearerAuth: []
      summary: Get All Article
      tags:
      - CMS - Article
    post:
      parameters:
      - description: Article payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ArticleCreateRequestDto'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Create Article
      tags:
      - CMS - Article
  /newsroom/articles/{id}:
    delete:
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Delete Article
      tags:
      - CMS - Article
    get:
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ArticleDetailResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get Article Details
      tags:
      - CMS - Article
    patch:
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      - description: Article payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ArticleUpdateRequestDto'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Update Article
      tags:
      - CMS - Article
  /newsroom/auth/info:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserInfoResponse'
              type: object
      summary: Auth Info
      tags:
      - CMS - Auth
  /newsroom/auth/login:
    post:
      parameters:
      - description: Auth payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AuthRequestDto'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Auth Login
      tags:
      - CMS - Auth
securityDefinitions:
  XApiKey:
    in: header
    name: X-Api-Key
    type: apiKey
  XRequestTime:
    in: header
    name: X-Request-Time
    type: apiKey
  XTempoToken:
    in: header
    name: X-Tempo-Token
    type: apiKey
swagger: "2.0"